requirements.txt
django>=1.4.3
djangorestframework>=2.1.15
biplist
